Modification to compile with MinGW 13.1 as dll (#147)
* add export declspec via HWINFO_API

* explicit cast to avoid type conversion warning

* Disable warning and msvc pragmas, add linking for mingw

* Add dll export definition to friend declarations

+++ src/windows/battery.cpp
@@ -34,7 +34,7 @@ std::vector<Battery> getAllBatteries() {
   IWbemClassObject* obj = nullptr;
   int battery_id = 0;
   while (wmi.enumerator) {
-    wmi.enumerator->Next(WBEM_INFINITE, 1, &obj, &u_return);
+    wmi.enumerator->Next((long)WBEM_INFINITE, 1, &obj, &u_return);
     if (!u_return) {
       break;
     }
