+++ src/windows/mainboard.cpp
@@ -23,7 +23,7 @@ MainBoard::MainBoard() {
   }
   ULONG u_return = 0;
   IWbemClassObject* obj = nullptr;
-  wmi.enumerator->Next(WBEM_INFINITE, 1, &obj, &u_return);
+  wmi.enumerator->Next((long)WBEM_INFINITE, 1, &obj, &u_return);
   if (!u_return) {
     return;
   }
